% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_search.R
\name{stochastic_search}
\alias{stochastic_search}
\title{Stochastic Search}
\usage{
stochastic_search(dim_param, perf_fun, loss_fun = ls_loss, target_perf,
  max_iter = 100, tol = 0, curiosity = 1, block_num, block_size,
  lambda = 0, reg_fun = zero_reg, param)
}
\arguments{
\item{dim_param}{integer; the dimension of the parameters.}

\item{perf_fun}{function; the performance function mapping parameters to an outcome,
e.g. a curve.}

\item{loss_fun}{function; the loss function of two outcomes.}

\item{target_perf}{function; the performance to be matched, e.g. a target data curve.}

\item{max_iter}{integer; maximum number of iterations.}

\item{tol}{numeric; tolerence for the loss functions.}

\item{curiosity}{numeric; search parameter.}

\item{block_num}{integer.}

\item{block_size}{integer.}

\item{lambda}{numeric; regularisation parameter.}

\item{reg_fun}{function; regularisation funciton, see 'zero_reg', 'L1_reg', 'L2_reg'.}

\item{param}{numeric vector; parameter from previous run of the function.}
}
\description{
An stochasic algorithm for optimisation
}
\details{
Denote the dimension of the parameters by d.
If the stochastic search didn't improve in 5*d block iterations, then it will
increase the curiosity by 10 folds. On the other hand, all improvements are
associated with a 0.001 decrease curiosity. This is to create oscillatary
behaviour of the curiosity parameter.
}

