% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochasticSearch.R
\docType{package}
\name{stochasticSearch}
\alias{stochasticSearch}
\alias{stochasticSearch-package}
\title{A package for stochastic search.}
\description{
This package implements a basic stochastic search algorithm.
}
\author{
Jackson Kwok
}

