% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandeRer.R
\docType{package}
\name{wandeRer}
\alias{wandeRer}
\alias{wandeRer-package}
\title{A package for stochastic search.}
\description{
This package implements a basic block-based, self-correcting stochastic search algorithm.
}
\author{
Jackson Kwok
}

